requirements.txt
pyqt5
whisper
ffmpeg-python
srt